Downgraded Ubuntu from 22.04 to 18.04
The reason this was downgraded is because Ubuntu 22.04 requires at least
TLS version 1.2, while the WABO server only accepts TLS version 1.1. For
this reason, stunnel cannot setup the tunnel, and the deployment fails.
We now have a new endpoint for the pre-wabo/edepot server, and to test
this we downgrade this back to 18.04. We can later revert this commit to
upgrade back to 22.04 once that server is upgraded to accept TLS
version 1.2